Remove feature flag and  redundant spec files
