+++ spec/system/candidate_interface/course_selection/candidate_existing_user_with_course_params_spec.rb
@@ -4,7 +4,6 @@ RSpec.describe 'An existing candidate arriving from Find with a course and provi
   include CourseOptionHelpers
   scenario 'candidate is not signed in and retains their course selection through the sign in process' do
     given_the_pilot_is_open
-    and_course_selection_page_is_active
     and_the_create_account_or_sign_in_page_feature_flag_is_active
 
     # Single site course
@@ -59,10 +58,6 @@ RSpec.describe 'An existing candidate arriving from Find with a course and provi
 
   def given_the_pilot_is_open
     FeatureFlag.activate('pilot_open')
-  end
-
-  def and_course_selection_page_is_active
-    FeatureFlag.activate('you_selected_a_course_page')
   end
 
   def and_the_create_account_or_sign_in_page_feature_flag_is_active
