+++ spec/system/candidate_interface/course_selection/candidate_signed_in_user_with_course_params_spec.rb
@@ -4,7 +4,6 @@ RSpec.describe 'An existing candidate arriving from Find with a course and provi
   include CourseOptionHelpers
   scenario 'candidate is signed in' do
     given_the_pilot_is_open
-    and_course_selection_page_is_active
     and_i_am_an_existing_candidate_on_apply
     and_i_have_less_than_3_application_options
     and_the_course_i_selected_only_has_one_site
@@ -50,10 +49,6 @@ RSpec.describe 'An existing candidate arriving from Find with a course and provi
 
   def given_the_pilot_is_open
     FeatureFlag.activate('pilot_open')
-  end
-
-  def and_course_selection_page_is_active
-    FeatureFlag.activate('you_selected_a_course_page')
   end
 
   def and_the_course_i_selected_only_has_one_site
